move region to components and add base component class
