--- hydromt/models/region/__init__.py
@@ -1,5 +1,0 @@
-"""All functionality to do with defining a model region."""
-
-from hydromt.models.region.region import ModelRegion
-
-__all__ = ["ModelRegion"]
